tests: refactor payment service
