+++ src/app/modules/form/admin-form/__tests__/admin-form.service.spec.ts
@@ -28,7 +28,6 @@ import { EditFieldActions } from 'src/shared/constants'
 import {
   FormLogicSchema,
   IEmailFormSchema,
-  IEncryptedFormDocument,
   IFormDocument,
   IFormSchema,
   IPopulatedForm,
@@ -55,12 +54,10 @@ import {
   FormStatus,
   LogicDto,
   LogicType,
-  PaymentsUpdateDto,
   SettingsUpdateDto,
 } from '../../../../../../shared/types'
 import { smsConfig } from '../../../../config/features/sms.config'
 import * as SmsService from '../../../../services/sms/sms.service'
-import { InvalidPaymentAmountError } from '../../../payments/payments.errors'
 import {
   FormNotFoundError,
   LogicNotFoundError,
@@ -2676,121 +2673,6 @@ describe('admin-form.service', () => {
       expect(actualResult._unsafeUnwrap()).toEqual(1)
 
       expect(twilioCacheSpy).toHaveBeenCalledWith(MSG_SRVC_NAME)
-    })
-  })
-
-  describe('updatePayments', () => {
-    // Arrange
-    const mockFormId = new ObjectId().toString()
-    const updatedPaymentSettings: PaymentsUpdateDto = {
-      enabled: true,
-      amount_cents: 5000,
-      description: 'some description',
-    }
-
-    const mockUpdatedForm = {
-      _id: mockFormId,
-      payments_field: updatedPaymentSettings,
-    }
-
-    it('should return InvalidPaymentAmountError if payment amount exceeds maxPaymentAmountCents', async () => {
-      // Arrange
-
-      const updatedPaymentSettingsExceeded = {
-        ...updatedPaymentSettings,
-        amount_cents: 100000001,
-      } as PaymentsUpdateDto
-
-      // Act
-      const actualResult = await AdminFormService.updatePayments(
-        mockFormId,
-        updatedPaymentSettingsExceeded,
-      )
-
-      // Assert
-      expect(actualResult.isErr()).toEqual(true)
-      expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(
-        InvalidPaymentAmountError,
-      )
-    })
-
-    it('should return InvalidPaymentAmountError if payment amount is below minPaymentAmountCents', async () => {
-      // Arrange
-
-      const updatedPaymentSettingsBelow = {
-        ...updatedPaymentSettings,
-        amount_cents: 49,
-      } as PaymentsUpdateDto
-
-      // Act
-      const actualResult = await AdminFormService.updatePayments(
-        mockFormId,
-        updatedPaymentSettingsBelow,
-      )
-
-      // Assert
-      expect(actualResult.isErr()).toEqual(true)
-      expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(
-        InvalidPaymentAmountError,
-      )
-    })
-
-    it('should successfuly call updatePaymentsById with formId and newPayments and return the updated payment settings', async () => {
-      // Arrange
-      const putSpy = jest
-        .spyOn(EncryptFormModel, 'updatePaymentsById')
-        .mockResolvedValueOnce(
-          mockUpdatedForm as unknown as IEncryptedFormDocument,
-        )
-
-      // Act
-      const actualResult = await AdminFormService.updatePayments(
-        mockFormId,
-        updatedPaymentSettings,
-      )
-
-      // Assert
-      expect(putSpy).toHaveBeenCalledWith(mockFormId, updatedPaymentSettings)
-
-      expect(actualResult.isOk()).toEqual(true)
-      // Should equal updatedPaymentSettings obj
-      expect(actualResult._unsafeUnwrap()).toEqual(updatedPaymentSettings)
-    })
-
-    it('should return PossibleDatabaseError if db update fails', async () => {
-      // Arrange
-      const putSpy = jest
-        .spyOn(EncryptFormModel, 'updatePaymentsById')
-        .mockRejectedValueOnce(new DatabaseError())
-
-      // Act
-      const actualResult = await AdminFormService.updatePayments(
-        mockFormId,
-        updatedPaymentSettings,
-      )
-
-      // Assert
-      expect(putSpy).toHaveBeenCalledWith(mockFormId, updatedPaymentSettings)
-      expect(actualResult.isErr()).toEqual(true)
-      expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(DatabaseError)
-    })
-
-    it('should return FormNotFoundError if no form is returned after updating db', async () => {
-      // Arrange
-      const putSpy = jest
-        .spyOn(EncryptFormModel, 'updatePaymentsById')
-        .mockResolvedValueOnce(null)
-
-      // Act
-      const actualResult = await AdminFormService.updatePayments(
-        mockFormId,
-        updatedPaymentSettings,
-      )
-
-      // Assert
-      expect(putSpy).toHaveBeenCalledWith(mockFormId, updatedPaymentSettings)
-      expect(actualResult.isErr()).toEqual(true)
-      expect(actualResult._unsafeUnwrapErr()).toBeInstanceOf(FormNotFoundError)
     })
   })
 })
