+++ src/app/modules/form/admin-form/admin-form.service.ts
@@ -28,8 +28,6 @@ import {
   FormSettings,
   LogicDto,
   MobileFieldBase,
-  PaymentsUpdateDto,
-  PaymentType,
   SettingsUpdateDto,
   StartPageUpdateDto,
 } from '../../../../../shared/types'
@@ -37,7 +35,6 @@ import { EditFieldActions } from '../../../../shared/constants'
 import {
   FormFieldSchema,
   FormLogicSchema,
-  IEncryptedFormDocument,
   IForm,
   IFormDocument,
   IFormSchema,
@@ -45,7 +42,6 @@ import {
 } from '../../../../types'
 import { EditFormFieldParams, FormUpdateParams } from '../../../../types/api'
 import config, { aws as AwsConfig } from '../../../config/config'
-import { paymentConfig } from '../../../config/features/payment.config'
 import { createLoggerWithLabel } from '../../../config/logger'
 import getAgencyModel from '../../../models/agency.server.model'
 import getFormModel, {
@@ -69,7 +65,6 @@ import {
   SecretsManagerNotFoundError,
   TwilioCacheError,
 } from '../../core/core.errors'
-import { InvalidPaymentAmountError } from '../../payments/payments.errors'
 import { MissingUserError } from '../../user/user.errors'
 import * as UserService from '../../user/user.service'
 import { SmsLimitExceededError } from '../../verification/verification.errors'
@@ -106,9 +101,9 @@ import {
   processDuplicateOverrideProps,
 } from './admin-form.utils'
 
-const logger = createLoggerWithLabel(module)
+export const logger = createLoggerWithLabel(module)
 const FormModel = getFormModel(mongoose)
-const EncryptedFormModel = getEncryptedFormModel(mongoose)
+export const EncryptedFormModel = getEncryptedFormModel(mongoose)
 const AgencyModel = getAgencyModel(mongoose)
 
 export const secretsManager = new SecretsManager({
@@ -1571,61 +1566,6 @@ const deleteTwilioTransaction = async (
     })
     throw new SecretsManagerError(awsError.message)
   }
-}
-
-/**
- * Update the payments field of the given form
- * @param formId the id of the form to update the end page for
- * @param newPayments the new payments field to replace the current one
- * @returns ok(updated payments object) when update is successful
- * @returns err(FormNotFoundError) if form cannot be found
- * @returns err(PossibleDatabaseError) if start page update fails
- * @returns err(InvalidPaymentAmountError) if payment amount exceeds MAX_PAYMENT_AMOUNT
- */
-export const updatePayments = (
-  formId: string,
-  newPayments: PaymentsUpdateDto,
-): ResultAsync<
-  IEncryptedFormDocument['payments_field'],
-  PossibleDatabaseError | FormNotFoundError | InvalidPaymentAmountError
-> => {
-  const { enabled } = newPayments
-
-  // Check if payment amount exceeds maxPaymentAmountCents or below minPaymentAmountCents if the payment is enabled
-  if (enabled && newPayments?.payment_type === PaymentType.Fixed) {
-    const { amount_cents } = newPayments
-    if (
-      amount_cents > paymentConfig.maxPaymentAmountCents ||
-      amount_cents < paymentConfig.minPaymentAmountCents
-    ) {
-      return errAsync(new InvalidPaymentAmountError())
-    }
-  }
-
-  return ResultAsync.fromPromise(
-    EncryptedFormModel.updatePaymentsById(formId, newPayments),
-    (error) => {
-      logger.error({
-        message: 'Error occurred when updating form payments',
-        meta: {
-          action: 'updatePayments',
-          formId,
-          newPayments,
-        },
-        error,
-      })
-      return transformMongoError(error)
-    },
-  ).andThen((updatedForm) => {
-    if (!updatedForm) {
-      return errAsync(new FormNotFoundError())
-    }
-    return okAsync(updatedForm.payments_field)
-  })
-}
-
-export const getPaymentGuideLink = (): string => {
-  return paymentConfig.guideLink
 }
 
 export const getGoLinkSuffix = (formId: string) => {
